protect the DB via Configuration flag
git-svn-id: http://subversion.assembla.com/svn/WebAlbums@212 373d8ef0-abc8-4151-9680-7fa2cb6b1372

+++ WebAlbums3/trunk/WebAlbums3-DAO-JPABeans/src/java/net/wazari/dao/jpa/ThemeFacade.java
@@ -27,7 +27,10 @@ public class ThemeFacade implements ThemeFacadeLocal {
     private EntityManager em;
 
     @Override
-    public void remove(Theme theme) {
+    public void remove(Theme theme, boolean protect) {
+        if (protect || WebAlbumsDAOBean.PERSISTENCE_UNIT != WebAlbumsDAOBean.PERSISTENCE_UNIT_MySQL) {
+            throw new IllegalStateException("cannot remove a theme while ProtectedDB is enabled") ;
+        }
         em.remove(em.merge(theme));
     }
